requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-jazzmin==3.0.1
dnspython==2.8.0
email-validator==2.3.0
frozenlist==1.8.0
idna==3.11
mercadopago==2.3.0
multidict==6.7.1
pillow==12.1.0
propcache==0.4.1
PyJWT==2.11.0
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
twilio==9.10.1
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
